rename and restructure levels and settings
