+++ src/app/app.routes.ts
@@ -1,9 +1,9 @@
 import { Routes } from '@angular/router';
-import { Level1pageComponent } from './level-1-page/level-1-page.component'
+import { Level1pageComponent } from './levels/level-1-page/level-1-page.component'
 import { SettingsComponent } from './settings/settings.component'
 import { LearnComponent } from './learn/learn.component';
 import { HomeComponent } from './home/home.component';
-import { Level2PageComponent } from './level-2-page/level-2-page.component';
+import { Level2PageComponent } from './levels/level-2-page/level-2-page.component';
 import { LearnRoutingService } from "./learn/routing.service";
 
 const routingService = new LearnRoutingService();
