+++ src/app/levels/buttons-overlay/buttons-overlay.component.scss
@@ -1,4 +1,4 @@
-@import '../../_variables';
+@import '../../../_variables';
 
 .overlay {
