+++ src/app/levels/buttons-overlay/buttons-overlay.component.ts
