+++ src/app/levels/level-1-page/level-1-page.component.scss
@@ -1,4 +1,4 @@
-@import '../../variables';
+@import '../../../variables';
 
 
 .level-info, #level-1-quiz-wrapper {
