+++ src/app/levels/level-1-quiz-question/level-1-quiz-question.component.scss
@@ -1,4 +1,4 @@
-@import '../../variables';
+@import '../../../variables';
 
 .level-1-quiz-question-wrapper {
   display: flex;
