+++ src/app/levels/level-2-page/level-2-page.component.scss
