+++ src/app/levels/level-2-quiz-question/level-2-quiz-question.component.scss
@@ -1,4 +1,4 @@
-@import '../../variables';
+@import '../../../variables';
 
 $border-radius: 0.625rem;
