+++ src/app/levels/on-screen-keyboard/on-screen-keyboard.component.scss
@@ -1,4 +1,4 @@
-@import '../../variables';
+@import '../../../variables';
 
 * {
  margin: 0;
