+++ src/app/levels/on-screen-keyboard/on-screen-keyboard.component.spec.ts
