+++ src/app/levels/on-screen-keyboard/on-screen-keyboard.component.ts
