+++ src/app/settings/hue-selector/hue-selector.component.html
