+++ src/app/settings/hue-selector/hue-selector.component.scss
@@ -1,4 +1,4 @@
-@import "../../variables";
+@import "../../../variables";
 
 .hue-selector {
