+++ src/app/settings/segmented-multiple-selection/segmented-multiple-selection.component.html
