+++ src/app/settings/segmented-multiple-selection/segmented-multiple-selection.component.scss
@@ -1,4 +1,4 @@
-@import "../../variables";
+@import "../../../variables";
 
 .segmented-multiple-selection {
