+++ src/app/settings/segmented-single-selection/segmented-single-selection.component.scss
@@ -1,4 +1,4 @@
-@import "../../variables";
+@import "../../../variables";
 
 .segmented-single-selection {
