+++ src/app/settings/segmented-single-selection/segmented-single-selection.component.spec.ts
