+++ src/app/settings/settings.component.ts
@@ -1,8 +1,8 @@
 import { Component } from '@angular/core';
-import { SegmentedSingleSelectionComponent } from '../segmented-single-selection/segmented-single-selection.component';
-import { SegmentedMultipleSelectionComponent } from '../segmented-multiple-selection/segmented-multiple-selection.component';
-import { ThemeSelectorComponent } from '../theme-selector/theme-selector.component';
-import { HueSelectorComponent } from '../hue-selector/hue-selector.component';
+import { SegmentedSingleSelectionComponent } from './segmented-single-selection/segmented-single-selection.component';
+import { SegmentedMultipleSelectionComponent } from './segmented-multiple-selection/segmented-multiple-selection.component';
+import { ThemeSelectorComponent } from './theme-selector/theme-selector.component';
+import { HueSelectorComponent } from './hue-selector/hue-selector.component';
 
 @Component({
     selector: 'app-settings',
