+++ src/app/settings/theme-selector/theme-selector.component.scss
@@ -1,7 +1,7 @@
 @use "sass:map";
 @use "sass:color";
 
-@import '../../variables';
+@import "../../../variables";
 
 .theme-selector {
