+++ src/app/settings/theme-selector/theme-selector.component.ts
