+++ src/app/settings/theme.service.spec.ts
